apps/api/requirements.txt
azure-functions>=1.17.0
azure-cosmos>=4.5.1
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
openai>=1.12.0
httpx>=0.27.0
amadeus>=9.0.0
stripe>=8.0.0
pydantic>=2.5.0
python-dateutil>=2.8.2
qrcode>=7.4.2
Pillow>=10.2.0
reportlab>=4.1.0
